backend/pyproject.toml
[project]
name = "dia-backend"
version = "0.1.0"
description = "Decision Intelligence Assistant — FastAPI backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn[standard]==0.34.2",
    "pydantic==2.11.4",
    "pydantic-settings==2.9.1",
    "qdrant-client==1.14.2",
    "ollama==0.4.8",
    "google-generativeai==0.8.5",
    "scikit-learn==1.6.1",
    "joblib==1.4.2",
    "pandas==2.2.3",
    "numpy==2.2.5",
    "python-dotenv==1.1.0",
]

[dependency-groups]
dev = [
    "pytest==8.3.5",
    "pytest-cov==6.1.0",
    "pytest-asyncio==0.25.3",
    "httpx==0.28.1",
    "black==24.10.0",
    "isort==5.13.2",
    "flake8==7.1.1",
    "mypy==1.13.0",
    "types-requests==2.32.0.20241016",
]

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
strict = true
ignore_missing_imports = true
python_version = "3.13"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
# cov-fail-under enforced once all service implementations are complete (Phase 2+)
addopts = "--cov=app --cov-report=term-missing"
